summary backend

+++ quiz/views.py
@@ -1,4 +1,5 @@
 from datetime import datetime, timezone
+from pdb import line_prefix
 from django.shortcuts import redirect, render, get_object_or_404
 from django.http import Http404
 
@@ -8,6 +9,7 @@ from django.contrib.auth.models import User
 
 from django.contrib import messages
 
+from django.db.models import OuterRef, Subquery
 from .models import Choice, ChoicePerUser, Question, Test, Test_event, Test_user
 
 # Create your views here.
@@ -45,21 +47,22 @@ def quiz_logic(request, quiz_id, question_nb):
         )
     except Test.DoesNotExist:
         raise Http404("Quiz does not exist")
-
+    print(user_picks)
     if quiz.test_event.start_date < datetime.now(timezone.utc):
         if request.method == "POST":
             print("POST")
-            ChoicePerUser.objects.all().filter(
-                test_user=quiz.id, pick__in=choices
-            ).delete()
-            if "ans" in request.POST:
-                picked = choices.filter(pk__in=request.POST.getlist("ans"))
-                print(datetime.today())
-                for i in range(len(picked)):
-                    picked[i].picks.add(
-                        quiz.id,
-                        through_defaults={"ans_time": datetime.today()},
-                    )
+            if quiz.test_event.end_date >= datetime.now(timezone.utc) or not user_picks:
+                ChoicePerUser.objects.all().filter(
+                    test_user=quiz.id, pick__in=choices
+                ).delete()
+                if "ans" in request.POST:
+                    picked = choices.filter(pk__in=request.POST.getlist("ans"))
+                    print(datetime.today())
+                    for i in range(len(picked)):
+                        picked[i].picks.add(
+                            quiz.id,
+                            through_defaults={"ans_time": datetime.today()},
+                        )
 
             if "DONE" in request.POST:
                 return redirect("summary", quiz_id=quiz_id)
@@ -88,24 +91,55 @@ def quiz_logic(request, quiz_id, question_nb):
 @login_required
 def summary(request, quiz_id):
     quiz = get_object_or_404(Test_user, user=request.user.id, test_event=quiz_id)
-    quiz.taken = True
-    quiz.score = 0
-    # calculate score #TODO
-    # questions = Question.objects.all().filter(test=quiz.test_event.test.id)
-    # print(questions)
-    # for q in questions:
-    #     choices = Choice.objects.all().filter(question_id__in=q)
-    #     print(choices)
+    questions = Question.objects.all().filter(test=quiz.test_event.test.id)
+    all_choices = Choice.objects.all().filter(question_id__in=questions.values("id"))
+    all_user_pick = ChoicePerUser.objects.all().filter(
+        pick__in=all_choices.values("id"), test_user=quiz.id
+    )
 
-    #     user_picks = (
-    #         ChoicePerUser.objects.all()
-    #         .filter(test_user=quiz.id, pick__in=choices)
-    #         .values_list("pick", flat=True)
-    #     )
-    #     print(user_picks)
+    first_date = Subquery(
+        ChoicePerUser.objects.all()
+        .filter(
+            test_user=quiz.id,
+            pick__question=OuterRef("question_id"),
+        )
+        .values("ans_time")[:1]
+    )
+    all_choices = all_choices.annotate(ans_time=first_date)
 
-    quiz.save()
-    return render(request, "quiz/index.html")
+    # if test end calculate score
+    if not quiz.taken:
+        # if True:
+        quiz.taken = True
+        quiz.score = 0
+        for i in range(len(questions)):
+            choices = all_choices.filter(question_id=questions[i].id)
+            user_picks = all_user_pick.all().filter(
+                test_user=quiz.id, pick__in=choices.values("id")
+            )
+            right_choices = choices.filter(if_correct=True)
+            # print(set(right_choices.values_list("id", flat=True)))
+            # print(set(user_picks.values_list("pick", flat=True)))
+            # print(user_picks.first().ans_time <= quiz.test_event.end_date)
+            if (
+                set(right_choices.values_list("id", flat=True))
+                == set(user_picks.values_list("pick", flat=True))
+                and user_picks.first().ans_time <= quiz.test_event.end_date
+            ):
+
+                quiz.score += 1
+        quiz.save()
+    return render(
+        request,
+        "quiz/summary.html",
+        {
+            "quiz": quiz,
+            "questions": questions,
+            "all_choices": all_choices,
+            "all_user_pick": set(all_user_pick.values_list("pick", flat=True)),
+            "test_event": quiz.test_event,
+        },
+    )
 
 
 def login_page(request):
